fix: resolve cross-platform CI failures
- shm.rs: cast mode_t to c_uint for macOS variadic shm_open (E0617)
- ci.yml: drop --all-targets from test run so --test-threads is not
  forwarded to criterion benches (invalid argument on all platforms)

+++ core/src/sink/shm.rs
@@ -164,7 +164,7 @@ mod shm_platform {
             libc::shm_open(
                 name_c.as_ptr(),
                 libc::O_CREAT | libc::O_EXCL | libc::O_RDWR,
-                permissions as libc::mode_t,
+                permissions as libc::mode_t as libc::c_uint,
             )
         };
         if fd < 0 {
